oh boy

+++ install.sh
@@ -4,8 +4,12 @@ CWD="$(pwd)"
 HOME_DIR="$HOME"
 CONFIG_DIR="$HOME_DIR/.config"
 
-echo $CONFIG_DIR
-
 ln -s "$CWD/tmux.conf" "$HOME_DIR/.tmux.conf"
 ln -s "$CWD/vimrc" "$HOME_DIR/.vimrc"
-ln -s "$CWD/nvim/" "$CONFIG_DIR/nvim"
+ln -s "$CWD/bashrc" "$HOME_DIR/.bashrc"
+
+if [ ! -f "$CONFIG_DIR/nvim" ]; then
+    ln -s "$CWD/nvim/" "$CONFIG_DIR/nvim"
+else
+    echo "Nvim is already there man!!!"
+fi
